# apps/java-governance/src/main/java/com/cosmic/governance/api/dto/JobStatusResponse.java
package com.cosmic.governance.api.dto;

import java.util.Map;

public record JobStatusResponse(
        String jobId,
        String workflow,
        String datasetId,
        String status,
        String createdAt,
        String updatedAt,
        Map<String, Object> parameters,
        String requestedBy
) {}